app/requirements.txt
numpy>=1.16.1
pytz>=2018.9
pytest>=4.2.0
flake8>=3.7.5
pytest-cov>=2.6.1
pylint>=2.3.1
mypy>=0.660
pandas>=0.24.1
alembic>=1.0.7
adal==1.2.1
requests>=2.21.0
msrestazure==0.6.0
